Add Amount type (cents storage), dev seed data, enriched products API, customer timeline
Amount type:
- New Amount newtype (i64 cents) in src/amount.rs
- Serializes as float in JSON (÷100), deserializes from float (×100)
- SQLite INTEGER columns for all money fields (was REAL)
- Arithmetic helpers: mul_qty, Add, Sub, Sum
- All models and routes updated to use Amount

Enriched products API:
- GET /products now returns prices HashMap per customer group
- e.g. { "prices": { "retail": 63000.0, "reseller": 54000.0 } }
- Quote form uses this directly instead of separate price lookup

Customer timeline:
- Added payment events to timeline (quotes, sales, bookings, payments)
- Fixed Amount type for timeline event amounts
- Quote detail shows customer name

Dev seed data:
- 8 products, 5 services, 2 warehouses
- 4 retail + 3 reseller customers with Paraguayan names/addresses
- 3 inventory receipts with prices for both customer groups
- 3 work teams
- Only runs in debug mode (cargo run), skipped in release

+++ migrations/002_products_inventory.sql
@@ -51,7 +51,7 @@ CREATE TABLE sales (
     customer_id       INTEGER NOT NULL REFERENCES customers(id),
     customer_group_id INTEGER NOT NULL REFERENCES customer_groups(id),
     notes             TEXT,
-    total_amount      REAL NOT NULL DEFAULT 0,
+    total_amount      INTEGER NOT NULL DEFAULT 0,
     sold_at           TEXT NOT NULL DEFAULT (datetime('now')),
     created_at        TEXT NOT NULL DEFAULT (datetime('now'))
 );
@@ -64,7 +64,7 @@ CREATE TABLE inventory_utxos (
     product_id       INTEGER NOT NULL REFERENCES products(id),
     warehouse_id     INTEGER NOT NULL REFERENCES warehouses(id),
     quantity         REAL NOT NULL CHECK (quantity > 0),
-    cost_per_unit    REAL NOT NULL CHECK (cost_per_unit >= 0),
+    cost_per_unit    INTEGER NOT NULL CHECK (cost_per_unit >= 0),
 
     -- Origin
     receipt_id       INTEGER REFERENCES inventory_receipts(id),
@@ -87,7 +87,7 @@ CREATE TABLE inventory_receipt_prices (
     receipt_id        INTEGER NOT NULL REFERENCES inventory_receipts(id),
     product_id        INTEGER NOT NULL REFERENCES products(id),
     customer_group_id INTEGER NOT NULL REFERENCES customer_groups(id),
-    price_per_unit    REAL NOT NULL CHECK (price_per_unit >= 0),
+    price_per_unit    INTEGER NOT NULL CHECK (price_per_unit >= 0),
     UNIQUE(receipt_id, product_id, customer_group_id)
 );
 
@@ -97,7 +97,7 @@ CREATE TABLE sale_lines (
     sale_id        INTEGER NOT NULL REFERENCES sales(id),
     product_id     INTEGER NOT NULL REFERENCES products(id),
     quantity       REAL NOT NULL CHECK (quantity > 0),
-    price_per_unit REAL NOT NULL,
+    price_per_unit INTEGER NOT NULL,
     created_at     TEXT NOT NULL DEFAULT (datetime('now'))
 );
